+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Fix color smoothing of elongated structures (by fixing ``Sphere.expand`` for spheres with highly directional extrema)
 - Fix entity label not displayed when multiple instances of the same entity are highlighted
 - Fix empty elements created in ``StructureElement.Loci.extendToAllInstances``
+- Measurement options tweaks (allow larger ``textSize``; make ``customText`` essential)
 
 ## [v3.0.1] - 2022-01-27